Alternative web path resolver added
