+++ Imagine/Cache/Resolver/AlternativeWebPathResolver.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Liip\ImagineBundle\Imagine\Cache\Resolver;
+
+class AlternativeWebPathResolver extends AbstractWebPathResolver
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function resolve($path, $filter)
+    {
+        return sprintf('/%s', $this->getFileUrl($path, $filter));
+    }
+}
